настроил контейнирезацию + теперь можно скачать список покупок

+++ backend/foodgram/recipes/models.py
@@ -50,7 +50,7 @@ class Recipe(models.Model):
 class RecipeIngredient(models.Model):
     recipe = models.ForeignKey(Recipe, on_delete=models.CASCADE, related_name='recipes_ingredient', verbose_name='Рецепт')
     ingredient = models.ForeignKey(Ingredient, on_delete=models.CASCADE, related_name='ingredients_recipe', verbose_name='Ингредиент')
-    amount = models.DecimalField(max_digits=10, decimal_places=2, verbose_name='Количество')
+    amount = models.FloatField(verbose_name='Количество')
 
     class Meta:
         constraints = (
